refactor(tools): separate system-related functions from tools.utils
KAG-2954
