+++ kong-3.6.0-0.rockspec
@@ -171,6 +171,7 @@ build = {
     ["kong.tools.yield"] = "kong/tools/yield.lua",
     ["kong.tools.uuid"] = "kong/tools/uuid.lua",
     ["kong.tools.rand"] = "kong/tools/rand.lua",
+    ["kong.tools.system"] = "kong/tools/system.lua",
 
     ["kong.runloop.handler"] = "kong/runloop/handler.lua",
     ["kong.runloop.events"] = "kong/runloop/events.lua",
